feat: add reusable GameRow component
